+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Demo case and demo sample loaded with demo instance startup
 - Endpoint for coverage queries over the intervals of a BED file
 - Include a default .env file loaded on app startup
+- Filter genes by Ensemble id, HGNC id, HGNC symbol
 
 ### Fixed